+++ alembic/versions/3bb3cd14b2dd_initial_version.py
@@ -5,7 +5,6 @@ Revises:
 Create Date: 2022-03-18 16:41:17.505723
 """
 import logging
-import typing
 
 # revision identifiers, used by Alembic.
 revision = "3bb3cd14b2dd"
@@ -14,9 +13,9 @@ branch_labels = None
 depends_on = None
 
 
-def upgrade(log: logging.Logger, table_names: typing.Set[str]) -> None:
+def upgrade(log: logging.Logger, table_names: set[str]) -> None:
     pass
 
 
-def downgrade(log: logging.Logger, table_names: typing.Set[str]) -> None:
+def downgrade(log: logging.Logger, table_names: set[str]) -> None:
     pass